# [perf] Improve ExportMap.for performance on larger codebases