@leipert 2023-04-08
This change is quite similar to #2755, but might be a little controversial, because it introduces a new dependency.

I had a look at `fast-deep-equal`: http://unpkg.com/browse/fast-deep-equal@3.1.3/index.js and I think it's simplicity aligns well with our use case here.